requirements.txt
pyaudio
cocos2d
